Move test to appropriate package
