+++ packages/forms/src/components/ui/text-input/__tests__/TextInput.test.tsx
@@ -1,19 +1,7 @@
 import { render } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import * as React from "react";
-import { useGridCell } from "@stenajs-webui/grid";
-import { TextInput, TextInputProps } from "../TextInput";
-
-const TextInputWithGridCell: React.FC<TextInputProps> = (props) => {
-  const { requiredProps } = useGridCell(props.value, {
-    rowIndex: 0,
-    colIndex: 0,
-    numRows: 10,
-    numCols: 10,
-    tableId: "test123",
-  });
-  return <TextInput {...requiredProps} {...props} />;
-};
+import { TextInput } from "../TextInput";
 
 describe("TextInput", () => {
   const ariaLabel = "Some input";
@@ -61,36 +49,6 @@ describe("TextInput", () => {
         expect(setValueMock).toBeCalledWith("5");
         expect(setValueMock).toBeCalledWith("6");
         expect(setValueMock).toBeCalledWith("7");
-      });
-    });
-
-    describe("and useGridCell is used", () => {
-      describe("gets called when text is entered", () => {
-        it("works", async () => {
-          const setValueMock = jest.fn();
-          const { getByLabelText } = render(
-            <TextInputWithGridCell
-              value={""}
-              onValueChange={setValueMock}
-              onMove={() => {}}
-              aria-label={ariaLabel}
-            />
-          );
-
-          const input = getByLabelText(ariaLabel);
-
-          await userEvent.type(input, "12");
-          await userEvent.type(input, "34567");
-
-          expect(setValueMock).toBeCalledTimes(7);
-          expect(setValueMock).toBeCalledWith("1");
-          expect(setValueMock).toBeCalledWith("2");
-          expect(setValueMock).toBeCalledWith("3");
-          expect(setValueMock).toBeCalledWith("4");
-          expect(setValueMock).toBeCalledWith("5");
-          expect(setValueMock).toBeCalledWith("6");
-          expect(setValueMock).toBeCalledWith("7");
-        });
       });
     });
